pyproject.toml
[project]
name = "job-recommendation-langchain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "apify-client>=2.5.0",
    "asyncio>=4.0.0",
    "dotenv>=0.9.9",
    "fastmcp>=3.1.1",
    "langchain>=1.2.12",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.1",
    "langchain-mcp-adapters>=0.2.2",
    "langgraph>=1.1.3",
    "pypdf>=6.9.1",
    "streamlit>=1.55.0",
]
